+++ CHANGELOG.md
@@ -294,3 +294,20 @@
 
                 - `item_detail_screen.dart` 중복 `DataManager` 지역 변수 제거
                     - `_onChatPressed`, `_onCancelPressed` 내부의 `DataManager dataManager = TestDataManager()` 중복 선언 제거
+
+- # feature/buildng
+    - ## version: 1.1.0
+        - ### commit: rental_item이 place사용하도록 구조 변경
+            - #### author: Seo JeongHun
+            - #### date: 2026-05-24
+            - feature:
+                - geofencing_api 패키지 설치
+                - LentalItem location -> placeID
+                - place.dart 수정
+                    - 필요없는 부분들 삭제
+                - building.dart 수정
+                    - 구조적으로 잘못된 부분 수정
+                    - 이웃 장소 불러오기 수정
+                - rental_request_screen.dart 수정
+                    - 장소를 드롭다운 형식으로 선택할 수 있게함
+                - test_data_manager에 건물 이름과 테두리 좌표 추가